requirements.txt
atomicwrites==1.4.1
attrs==22.1.0
black==22.10.0
certifi==2022.9.24
charset-normalizer==2.0.12
click==8.1.3
colorama==0.4.6
Django==2.2.16
django-behaviors==0.5.1
djhtml==1.5.2
Faker==12.0.1
flake8==5.0.4
flake8-commas==2.1.0
idna==3.4
importlib-metadata==4.2.0
iniconfig==1.1.1
isort==5.10.1
mccabe==0.7.0
mixer==7.1.2
mypy==0.990
mypy-extensions==0.4.3
packaging==21.3
pathspec==0.10.1
Pillow==8.3.1
platformdirs==2.5.3
pluggy==0.13.1
py==1.11.0
pycodestyle==2.9.1
pyflakes==2.5.0
pyparsing==3.0.9
pytest==6.2.4
pytest-django==4.4.0
pytest-pythonpath==0.7.3
python-dateutil==2.8.2
pytz==2022.6
requests==2.26.0
six==1.16.0
sorl-thumbnail==12.7.0
sqlparse==0.4.3
StrEnum==0.4.8
toml==0.10.2
tomli==2.0.1
typed-ast==1.5.4
typing_extensions==4.4.0
urllib3==1.26.12
zipp==3.10.0
